Move LinkButton functionality to utils
