--- src/utils/AppLinks.ts
@@ -1,3 +1,0 @@
-export enum Links {
-  GitHub = "https://github.com/yurixander/erc20-vault",
-}
