+++ src/utils/utils.ts
@@ -4,3 +4,7 @@ import { twMerge } from "tailwind-merge";
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }
+
+export enum Links {
+  GitHub = "https://github.com/yurixander/erc20-vault",
+}
